[pliron] Address review: NSuccsInterface + move op unit tests
Review feedback on PR #116:

- Add NSuccsInterface<const N>: BranchOpInterface and OneSuccInterface
  (providing get_successor()) to builtin op-interfaces, mirroring the
  operand-side NOpdsInterface / OneOpdInterface family. BrOp now implements
  OneSuccInterface (i.e. NSuccsInterface<1>) and CondBrOp NSuccsInterface<2>,
  replacing BrOp's hand-rolled single-successor Verify with the framework
  interface. The br / cond_br round-trips in compile_run.rs exercise it.

- Move the inline #[cfg(test)] mod tests out of ops.rs into
  tests/op_unit_tests.rs (alongside compile_run.rs). These are pure
  print/parse round-trips, so the file needs no llvm-sys feature gate.
