+++ pliron-llvm/src/ops.rs
@@ -11,9 +11,10 @@ use pliron::{
             self, ATTR_KEY_SYM_NAME, AtLeastNOpdsInterface, AtLeastNResultsInterface,
             AtMostNOpdsInterface, AtMostNRegionsInterface, AtMostOneRegionInterface,
             BranchOpInterface, CallOpCallable, CallOpInterface, IsTerminatorInterface,
-            IsolatedFromAboveInterface, NOpdsInterface, NResultsInterface, OneOpdInterface,
-            OneResultInterface, OperandNOfType, OperandSegmentInterface, OptionalOpdInterface,
-            ResultNOfType, SameOperandsAndResultType, SameOperandsType, SameResultsType,
+            IsolatedFromAboveInterface, NOpdsInterface, NResultsInterface, NSuccsInterface,
+            OneOpdInterface, OneResultInterface, OneSuccInterface, OperandNOfType,
+            OperandSegmentInterface, OptionalOpdInterface, ResultNOfType,
+            SameOperandsAndResultType, SameOperandsType, SameResultsType,
             SingleBlockRegionInterface, SymbolOpInterface, SymbolUserOpInterface,
         },
         type_interfaces::{FloatTypeInterface, FunctionTypeInterface},
@@ -682,26 +683,15 @@ pub struct AddrSpaceCastOp;
 #[pliron_op(
     name = "llvm.br",
     format = "succ($0) `(` operands(CharSpace(`,`)) `)`",
-    interfaces = [IsTerminatorInterface, NResultsInterface<0>],
+    interfaces = [
+        IsTerminatorInterface,
+        NResultsInterface<0>,
+        NSuccsInterface<1>,
+        OneSuccInterface
+    ],
+    verifier = "succ"
 )]
 pub struct BrOp;
-
-#[derive(Error, Debug)]
-enum BrOpVerifyErr {
-    #[error("BrOp must have exactly one successor")]
-    ExpectedOneSuccessor,
-}
-
-impl Verify for BrOp {
-    fn verify(&self, ctx: &Context) -> Result<()> {
-        // Operand/argument count and types are checked by BranchOpInterface;
-        // here we only enforce the single-successor invariant.
-        if self.get_operation().deref(ctx).get_num_successors() != 1 {
-            verify_err!(self.loc(ctx), BrOpVerifyErr::ExpectedOneSuccessor)?
-        }
-        Ok(())
-    }
-}
 
 #[op_interface_impl]
 impl BranchOpInterface for BrOp {
@@ -758,7 +748,7 @@ impl BrOp {
 /// | `false_dest` | Any successor |
 #[pliron_op(
     name = "llvm.cond_br",
-    interfaces = [IsTerminatorInterface, NResultsInterface<0>],
+    interfaces = [IsTerminatorInterface, NResultsInterface<0>, NSuccsInterface<2>],
 )]
 pub struct CondBrOp;
 impl CondBrOp {
@@ -4325,202 +4315,5 @@ impl Verify for FuncOp {
 impl IsDeclaration for FuncOp {
     fn is_declaration(&self, ctx: &Context) -> bool {
         self.get_region(ctx).is_none()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        AddrSpaceCastOp, AtomicCmpxchgOp, AtomicLoadOp, AtomicRmwOp, AtomicStoreOp, FenceOp,
-        FuncOp, InlineAsmOp, ReturnOp,
-    };
-    use crate::attributes::{AtomicOrderingAttr, AtomicRmwKindAttr};
-    use crate::op_interfaces::CastOpInterface;
-    use crate::types::{FuncType, PointerType, VoidType};
-    use pliron::{
-        basic_block::BasicBlock,
-        builtin::{
-            op_interfaces::SingleBlockRegionInterface,
-            ops::ModuleOp,
-            types::{IntegerType, Signedness},
-        },
-        combine::Parser,
-        context::{Context, Ptr},
-        irfmt::parsers::spaced,
-        location,
-        op::Op,
-        operation::Operation,
-        parsable::{self, state_stream_from_iterator},
-        printable::Printable,
-        r#type::TypeObj,
-        value::Value,
-    };
-
-    /// Build a module with a function whose entry block has the given argument
-    /// types, let `build` add ops to that block, then assert the whole module
-    /// round-trips through the parser unchanged.
-    fn assert_op_roundtrips(
-        make_arg_types: impl FnOnce(&mut Context) -> Vec<Ptr<TypeObj>>,
-        build: impl FnOnce(&mut Context, Ptr<BasicBlock>, &[Value]),
-        expected: &[&str],
-    ) {
-        let ctx = &mut Context::new();
-        let arg_types = make_arg_types(ctx);
-
-        let module = ModuleOp::new(ctx, "test".try_into().unwrap());
-        let module_block = module.get_body(ctx, 0);
-
-        let void = VoidType::get(ctx).into();
-        let fty = FuncType::get(ctx, void, arg_types, false);
-        let func = FuncOp::new(ctx, "f".try_into().unwrap(), fty);
-        let entry = func.get_or_create_entry_block(ctx);
-        let args: Vec<Value> = entry.deref(ctx).arguments().collect();
-
-        build(ctx, entry, &args);
-        ReturnOp::new(ctx, None)
-            .get_operation()
-            .insert_at_back(entry, ctx);
-        func.get_operation().insert_at_back(module_block, ctx);
-
-        let printed = module.get_operation().disp(ctx).to_string();
-        for tok in expected {
-            assert!(
-                printed.contains(tok),
-                "expected `{tok}` in printed IR:\n{printed}"
-            );
-        }
-
-        // The printed IR must re-parse: this exercises the op's (and its
-        // attributes') parser against exactly what its printer produced.
-        let ctx2 = &mut Context::new();
-        let state_stream = state_stream_from_iterator(
-            printed.chars(),
-            parsable::State::new(ctx2, location::Source::InMemory),
-        );
-        spaced(Operation::top_level_parser())
-            .parse(state_stream)
-            .expect("printed IR should re-parse");
-    }
-
-    fn i32_ty(ctx: &mut Context) -> Ptr<TypeObj> {
-        IntegerType::get(ctx, 32, Signedness::Signless).into()
-    }
-
-    #[test]
-    fn test_addrspacecast_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| vec![PointerType::get(ctx, 0).into()],
-            |ctx, entry, args| {
-                let dst = PointerType::get(ctx, 1).into();
-                AddrSpaceCastOp::new(ctx, args[0], dst)
-                    .get_operation()
-                    .insert_at_back(entry, ctx);
-            },
-            &["llvm.addrspacecast", "llvm.ptr addrspace(1)"],
-        );
-    }
-
-    #[test]
-    fn test_atomicrmw_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| vec![PointerType::get(ctx, 1).into(), i32_ty(ctx)],
-            |ctx, entry, args| {
-                AtomicRmwOp::new(
-                    ctx,
-                    args[0],
-                    args[1],
-                    AtomicRmwKindAttr::Add,
-                    AtomicOrderingAttr::SeqCst,
-                    Some("device".to_string()),
-                )
-                .get_operation()
-                .insert_at_back(entry, ctx);
-            },
-            &["llvm.atomicrmw", "Add", "SeqCst", "device"],
-        );
-    }
-
-    #[test]
-    fn test_cmpxchg_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| {
-                let i = i32_ty(ctx);
-                vec![PointerType::get(ctx, 0).into(), i, i]
-            },
-            |ctx, entry, args| {
-                AtomicCmpxchgOp::new(
-                    ctx,
-                    args[0],
-                    args[1],
-                    args[2],
-                    AtomicOrderingAttr::AcqRel,
-                    AtomicOrderingAttr::Monotonic,
-                    None,
-                )
-                .get_operation()
-                .insert_at_back(entry, ctx);
-            },
-            &["llvm.cmpxchg", "AcqRel", "Monotonic"],
-        );
-    }
-
-    #[test]
-    fn test_fence_roundtrips() {
-        assert_op_roundtrips(
-            |_ctx| vec![],
-            |ctx, entry, _args| {
-                FenceOp::new(ctx, AtomicOrderingAttr::SeqCst, Some("block".to_string()))
-                    .get_operation()
-                    .insert_at_back(entry, ctx);
-            },
-            &["llvm.fence", "SeqCst", "block"],
-        );
-    }
-
-    #[test]
-    fn test_atomic_load_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| vec![PointerType::get(ctx, 3).into()],
-            |ctx, entry, args| {
-                let res = i32_ty(ctx);
-                AtomicLoadOp::new(ctx, args[0], res, AtomicOrderingAttr::Acquire, None)
-                    .get_operation()
-                    .insert_at_back(entry, ctx);
-            },
-            &["llvm.atomic_load", "Acquire"],
-        );
-    }
-
-    #[test]
-    fn test_atomic_store_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| vec![i32_ty(ctx), PointerType::get(ctx, 3).into()],
-            |ctx, entry, args| {
-                AtomicStoreOp::new(
-                    ctx,
-                    args[0],
-                    args[1],
-                    AtomicOrderingAttr::Release,
-                    Some("device".to_string()),
-                )
-                .get_operation()
-                .insert_at_back(entry, ctx);
-            },
-            &["llvm.atomic_store", "Release", "device"],
-        );
-    }
-
-    #[test]
-    fn test_inline_asm_roundtrips() {
-        assert_op_roundtrips(
-            |ctx| vec![i32_ty(ctx)],
-            |ctx, entry, args| {
-                let res = i32_ty(ctx);
-                InlineAsmOp::new(ctx, res, vec![args[0]], "mov $0, $1", "=r,r", true)
-                    .get_operation()
-                    .insert_at_back(entry, ctx);
-            },
-            &["llvm.inline_asm", "mov", "=r,r", "convergent"],
-        );
     }
 }
